Group sequence alignment primitives
