+++ test/analysis/alignment/pairwise/test_alignment.py
@@ -1,12 +1,15 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Tests of character-level line alignment."""
+"""Tests of pairwise character alignment."""
 
 from __future__ import annotations
 
 import numpy as np
 
-from scinoephile.analysis.line_alignment import LineAlignment, LineAlignmentOperation
+from scinoephile.analysis.alignment.pairwise import (
+    LineAlignment,
+    LineAlignmentOperation,
+)
 from test.helpers import parametrize
